requirements.txt
Flask
numpy
opencv-python-headless
Pillow
pytesseract
